requirements.txt
matplotlib==3.8.4
networkx==3.3
numpy==2.0.0
python_Levenshtein==0.25.1
regex==2024.4.28
scikit_learn==1.4.2
tqdm==4.66.2
